packages/engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ralphite-engine"
version = "0.1.0"
description = "Local-first orchestration engine for Ralphite TUI"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.8,<3",
  "pyyaml>=6.0,<7",
  "ralphite-schemas"
]

[tool.uv.sources]
ralphite-schemas = { workspace = true }

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
